% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_student_data.R
\name{write_student_data}
\alias{write_student_data}
\title{Write Student Data}
\usage{
write_student_data(data, path = "student_data.csv")
}
\arguments{
\item{data}{A data frame containing student performance data.}

\item{path}{The file path to save the student data CSV file. Defaults to "student_data.csv".}
}
\description{
Writes student performance data to a specified CSV file.
}
